8_Building_Your_First_GenAI_Application_with_AWS_Data_Foundations/8.3_Building_Your_Life_Insurance_Agent/lambda/create-kb-extension/requirements.txt
# lambda/create-kb-extension/requirements.txt
boto3
psycopg2-binary
urllib3
sqlalchemy
